Move string_constraint functions out of string_refinement
